+++ src/ui/toolbar/menus/FileMenu.tsx
@@ -8,9 +8,9 @@ import RemoveProjectDialog from "../dialogs/RemoveProjectDialog";
 import { ProjectContext, ProjectLoadContext } from "../../contexts/ProjectContext";
 import OverwriteChangesAlert from "../dialogs/OverwriteChangesAlert";
 import SaveAsDialog from "../dialogs/SaveAsDialog";
-import { exportTemplate } from "../../../filesystem/templatestore";
 import ExportFontDialog from "../dialogs/ExportFontDialog";
 import ImportTemplateWarningDialog from "../dialogs/ImportTemplateWarningDialog";
+import ExportTemplateDialog from "../dialogs/ExportTemplateDialog";
 
 export default function FileMenu() {
     const [newProjectModalOpen, setIsNewProjectModalOpen] = useState(false)
@@ -19,6 +19,7 @@ export default function FileMenu() {
     const [overwriteAlertOpen, setIsOverwriteAlertOpen] = useState(false)
     const [overwriteAlertAcceptFunction, setOverwriteAlertAcceptFunction] = useState(() => () => {})
     const [saveAsModalOpen, setSaveAsModalOpen] = useState(false)
+    const [exportTemplateModalOpen, setExportTemplateModalOpen] = useState(false)
     const [importTemplateModalOpen, setImportTemplateModalOpen] = useState(false)
     const [exportFontModalOpen, setExportFontModalOpen] = useState(false)
 
@@ -86,18 +87,7 @@ export default function FileMenu() {
                 throw new Error("No project to export a template from")
             }
 
-            exportTemplate(project)
-
-            setProjectContext({
-                ...project,
-                lastExportSnapshot: {
-                    defaultCharacterWidth: project.defaultCharacterWidth,
-                    defaultCharacterHeight: project.defaultCharacterHeight,
-                    characterBase: project.characterBase,
-                    characterSet: project.characterSet,
-                },
-                dirty: true,
-            })
+            setExportTemplateModalOpen(true)
         } catch (e: any) {
             const toaster = await OverlayToaster.create({ position: "top-right" })
             toaster.show({
@@ -123,7 +113,7 @@ export default function FileMenu() {
                 <MenuItem key="save-project" icon="floppy-disk" text={`${project?.dirty ? "*" : ""}Save Project`} disabled={!project} onClick={saveCurrentProject} />
                 <MenuItem key="save-as" icon="folder-shared" text="Save As..." disabled={!project} onClick={() => setSaveAsModalOpen(true)} />
                 <MenuDivider />
-                <MenuItem key="export-template" icon="export" text="Export Template PNG" disabled={!project} onClick={exportCurrentTemplate} />
+                <MenuItem key="export-template" icon="export" text="Export Template" disabled={!project} onClick={exportCurrentTemplate} />
                 <MenuItem key="import-template" icon="import" text="Import Template PNG" disabled={!project} onClick={() => setImportTemplateModalOpen(true)} />
                 <MenuDivider />
                 <MenuItem key="export-font" icon="generate" text="Export Font" disabled={!project} onClick={() => setExportFontModalOpen(true)} />
@@ -133,6 +123,7 @@ export default function FileMenu() {
             <OpenProjectDialog isOpen={openProjectModalOpen} setIsOpen={setIsOpenProjectModalOpen} />
             <OverwriteChangesAlert isOpen={overwriteAlertOpen} setIsOpen={setIsOverwriteAlertOpen} onAccepted={overwriteAlertAcceptFunction} />
             <SaveAsDialog isOpen={saveAsModalOpen} setIsOpen={setSaveAsModalOpen} />
+            {project && <ExportTemplateDialog isOpen={exportTemplateModalOpen} setIsOpen={setExportTemplateModalOpen} project={project} />}
             <ImportTemplateWarningDialog isOpen={importTemplateModalOpen} setIsOpen={setImportTemplateModalOpen} />
             <ExportFontDialog isOpen={exportFontModalOpen} setIsOpen={setExportFontModalOpen} />
             <RemoveProjectDialog isOpen={removeProjectModalOpen} setIsOpen={setRemoveProjectModalOpen} />